+++ R/internal_utility.R
@@ -291,26 +291,11 @@
 #' @param j Value of function
 #' @return inverse of Poisson ICC function
 #' @noRd
-.findPoisVar <- function(j) {
+.findPoisVar <- function(icc, lambda) {
 
-  # 'declare' var
-  y <- NULL
+  l <- log( ( 1 / lambda ) + 1 )
+  return(icc/(1-icc) * l)
 
-  a <- seq(0, 20, by = 0.01)
-  dx <- data.table::data.table(a = a, y = exp(3 * a / 2) - exp(a / 2))
-
-  amin <- dx[y <= j][.N, a]
-
-  a <- seq(amin, amin + 1e-02, length = 101)
-  dx <- data.table::data.table(a = a, y = exp(3 * a / 2) - exp(a / 2))
-
-  amin <- dx[y <= j][.N, a]
-
-  a <- seq(amin, amin + 1e-04, length = 1001)
-  dx <- data.table::data.table(a = a, y = exp(3 * a / 2) - exp(a / 2))
-
-
-  return(dx[y <= j][.N, a])
 }
 
 #' Gamma distribution - variation to generate positive skew