+++ NEWS.md
@@ -1,8 +1,16 @@
 # simstudy (development version)
 
 ## New features
-* Added double-dot (dynamic) functionality to defSurv. Users can now specificy 
+* Added double-dot (dynamic) functionality to defSurv. Users can now specify 
 double-dot variables in scale, shape, and formula parameters.
+* It is possible to generate variable cluster sizes using the *clusterSize*
+distribution in `defData` and `defDataAdd`.
+
+## Major fix
+* Improved the random effect variance generation for function `iccRE` under the
+Poisson distribution. The current approach is based on the 2013 paper by
+Nakagawa & Schielzeth titled "A general and simple method for obtaining $R^2$ from
+generalized linear mixed-effects models"
 
 ## Minor fix
 * Modified internal function to speed up beta distribution data generation.